+++ ibis/backends/dask/execution/selection.py
@@ -33,7 +33,7 @@ from ..execution.util import (
 )
 
 
-@compute_projection.register(ir.ScalarExpr, ops.Selection, dd.DataFrame)
+@compute_projection.register(ir.Scalar, ops.Selection, dd.DataFrame)
 def compute_projection_scalar_expr(
     expr,
     parent,
@@ -67,7 +67,7 @@ def compute_projection_scalar_expr(
     return data.assign(**{name: scalar})[name]
 
 
-@compute_projection.register(ir.ColumnExpr, ops.Selection, dd.DataFrame)
+@compute_projection.register(ir.Column, ops.Selection, dd.DataFrame)
 def compute_projection_column_expr(
     expr,
     parent,
@@ -246,7 +246,7 @@ def _compute_predicates(
     Parameters
     ----------
     table_op : TableNode
-    predicates : List[ir.ColumnExpr]
+    predicates : List[ir.Column]
     data : pd.DataFrame
     scope : Scope
     timecontext: Optional[TimeContext]