+++ ibis/backends/datafusion/__init__.py
@@ -164,13 +164,13 @@ class Backend(BaseBackend):
             frame = self.compile(expr, params, **kwargs)
             table = _to_pyarrow_table(frame)
             return table.to_pandas()
-        elif isinstance(expr, ir.ColumnExpr):
+        elif isinstance(expr, ir.Column):
             # expression must be named for the projection
             expr = expr.name('tmp').to_projection()
             frame = self.compile(expr, params, **kwargs)
             table = _to_pyarrow_table(frame)
             return table['tmp'].to_pandas()
-        elif isinstance(expr, ir.ScalarExpr):
+        elif isinstance(expr, ir.Scalar):
             if expr.op().root_tables():
                 # there are associated datafusion tables so convert the expr
                 # to a selection which we can directly convert to a datafusion