+++ ibis/backends/impala/tests/test_udf.py
@@ -239,10 +239,10 @@ def test_mult_args(i32, d, s, b, t):
     )
 
     expr = func(i32, d, s, b, t)
-    assert issubclass(type(expr), ir.ColumnExpr)
+    assert issubclass(type(expr), ir.Column)
 
     expr = func(1, 1.0, 'a', True, ibis.timestamp('1961-04-10'))
-    assert issubclass(type(expr), ir.ScalarExpr)
+    assert issubclass(type(expr), ir.Scalar)
 
 
 def _register_udf(inputs, output, name):
@@ -344,12 +344,12 @@ def test_decimal_fail(udfcon, test_data_db, udf_ll):
     )
 
     expr = func(literal)
-    assert issubclass(type(expr), ir.ScalarExpr)
+    assert issubclass(type(expr), ir.Scalar)
     result = udfcon.execute(expr)
     assert result == Decimal(1)
 
     expr = func(col)
-    assert issubclass(type(expr), ir.ColumnExpr)
+    assert issubclass(type(expr), ir.Column)
     udfcon.execute(expr)
 
 
@@ -366,11 +366,11 @@ def test_mixed_inputs(udfcon, alltypes, test_data_db, udf_ll):
     )
 
     expr = func(alltypes.int_col, 1)
-    assert issubclass(type(expr), ir.ColumnExpr)
+    assert issubclass(type(expr), ir.Column)
     udfcon.execute(expr)
 
     expr = func(1, alltypes.int_col)
-    assert issubclass(type(expr), ir.ColumnExpr)
+    assert issubclass(type(expr), ir.Column)
     udfcon.execute(expr)
 
     expr = func(alltypes.int_col, alltypes.tinyint_col)
@@ -396,7 +396,7 @@ def identity_func_testing(
     )
 
     expr = func(literal)
-    assert issubclass(type(expr), ir.ScalarExpr)
+    assert issubclass(type(expr), ir.Scalar)
     result = udfcon.execute(expr)
     # Hacky
     if datatype == 'timestamp':
@@ -409,7 +409,7 @@ def identity_func_testing(
             np.testing.assert_allclose(result, udfcon.execute(literal), 5)
 
     expr = func(column)
-    assert issubclass(type(expr), ir.ColumnExpr)
+    assert issubclass(type(expr), ir.Column)
     udfcon.execute(expr)