+++ ibis/expr/analysis.py
@@ -279,7 +279,7 @@ class ExprSimplifier:
         lifted_node = type(node)(*lifted_args)
 
         result = type(expr)(lifted_node)
-        if isinstance(expr, ir.ValueExpr) and expr.has_name():
+        if isinstance(expr, ir.Value) and expr.has_name():
             result = result.name(expr.get_name())
 
         return result
@@ -551,9 +551,7 @@ def get_mutation_exprs(
                     else:
                         overwriting_cols_to_expr[name] = None
                     expr_contains_overwrite = True
-        elif (
-            isinstance(expr, ir.ValueExpr) and expr.get_name() in table_schema
-        ):
+        elif isinstance(expr, ir.Value) and expr.get_name() in table_schema:
             overwriting_cols_to_expr[expr.get_name()] = expr
             expr_contains_overwrite = True
 
@@ -705,7 +703,7 @@ class _PushdownValidate:
                 return lin.proceed, self._validate_column(expr)
             return lin.proceed, None
 
-        return lin.traverse(validate, expr, type=ir.ValueExpr)
+        return lin.traverse(validate, expr, type=ir.Value)
 
     def _validate_column(self, expr):
         if isinstance(self.parent, ops.Selection):
@@ -785,7 +783,7 @@ def windowize_function(expr, w=None):
         unchanged = True
         windowed_args = []
         for arg in op.args:
-            if not isinstance(arg, ir.ValueExpr):
+            if not isinstance(arg, ir.Value):
                 windowed_args.append(arg)
                 continue
 
@@ -1016,13 +1014,13 @@ class FilterValidator(ExprValidator):
         else:
             roots_valid = []
             for arg in op.flat_args():
-                if isinstance(arg, ir.ScalarExpr):
+                if isinstance(arg, ir.Scalar):
                     # arg_valid = True
                     pass
                 elif isinstance(arg, ir.TopKExpr):
                     # TopK not subjected to further analysis for now
                     roots_valid.append(True)
-                elif isinstance(arg, (ir.ColumnExpr, ir.AnalyticExpr)):
+                elif isinstance(arg, (ir.Column, ir.AnalyticExpr)):
                     roots_valid.append(self.shares_some_roots(arg))
                 elif isinstance(arg, ir.Expr):
                     raise NotImplementedError(repr((type(expr), type(arg))))
@@ -1177,7 +1175,7 @@ def is_reduction(expr):
             return True
 
         for arg in op.args:
-            if isinstance(arg, ir.ScalarExpr) and has_reduction(arg.op()):
+            if isinstance(arg, ir.Scalar) and has_reduction(arg.op()):
                 return True
 
         return False
@@ -1186,4 +1184,4 @@ def is_reduction(expr):
 
 
 def is_scalar_reduction(expr):
-    return isinstance(expr, ir.ScalarExpr) and is_reduction(expr)
+    return isinstance(expr, ir.Scalar) and is_reduction(expr)