+++ ibis/expr/rules.py
@@ -37,7 +37,7 @@ def highest_precedence_dtype(exprs):
 
     Parameters
     ----------
-    exprs : Iterable[ir.ValueExpr]
+    exprs : Iterable[ir.Value]
       A sequence of Expressions
 
     Returns
@@ -154,12 +154,12 @@ def value(dtype, arg, **kwargs):
 
 @validator
 def scalar(inner, arg, **kwargs):
-    return instance_of(ir.ScalarExpr, inner(arg, **kwargs))
+    return instance_of(ir.Scalar, inner(arg, **kwargs))
 
 
 @validator
 def column(inner, arg, **kwargs):
-    return instance_of(ir.ColumnExpr, inner(arg, **kwargs))
+    return instance_of(ir.Column, inner(arg, **kwargs))
 
 
 @validator
@@ -451,7 +451,7 @@ def is_computable_input(value, **kwargs):
 def named_literal(value, **kwargs):
     import ibis.expr.operations as ops
 
-    if not isinstance(value, ir.ScalarExpr):
+    if not isinstance(value, ir.Scalar):
         raise com.IbisTypeError(
             "`value` must be a scalar expression; "
             f"got value of type {type(value).__name__}"