+++ ibis/expr/types/generic.py
@@ -16,7 +16,7 @@ from .core import Expr, _binop
 
 
 @public
-class ValueExpr(Expr):
+class Value(Expr):
 
     """
     Base class for a data generating expression having a fixed and known type,
@@ -33,7 +33,7 @@ class ValueExpr(Expr):
 
         Returns
         -------
-        ValueExpr
+        Value
             `self` with name `name`
 
         Examples
@@ -67,7 +67,7 @@ class ValueExpr(Expr):
 
 
 @public
-class ScalarExpr(ValueExpr):
+class Scalar(Value):
     def to_projection(self):
         """
         Promote this column expression to a table projection
@@ -90,7 +90,7 @@ class ScalarExpr(ValueExpr):
 
 
 @public
-class ColumnExpr(ValueExpr):
+class Column(Value):
     def parent(self):
         return self._arg
 
@@ -124,7 +124,7 @@ class ColumnExpr(ValueExpr):
 
 
 @public
-class AnyValue(ValueExpr):
+class AnyValue(Value):
     def hash(self, how: str = "fnv") -> ir.IntegerValue:
         """Compute an integer hash value.
 
@@ -142,7 +142,7 @@ class AnyValue(ValueExpr):
 
         return ops.Hash(self, how).to_expr()
 
-    def cast(self, target_type: dt.DataType) -> ValueExpr:
+    def cast(self, target_type: dt.DataType) -> Value:
         """Cast expression to indicated data type.
 
         Parameters
@@ -152,7 +152,7 @@ class AnyValue(ValueExpr):
 
         Returns
         -------
-        ValueExpr
+        Value
             Casted expression
         """
         import ibis.expr.operations as ops
@@ -174,7 +174,7 @@ class AnyValue(ValueExpr):
             return result
         return result.name(f'cast({self.get_name()}, {op.to})')
 
-    def coalesce(self, *args: ValueExpr) -> ValueExpr:
+    def coalesce(self, *args: Value) -> Value:
         """Return the first non-null value from `args`.
 
         Parameters
@@ -184,7 +184,7 @@ class AnyValue(ValueExpr):
 
         Returns
         -------
-        ValueExpr
+        Value
             Coalesced expression
 
         Examples
@@ -198,7 +198,7 @@ class AnyValue(ValueExpr):
 
         return ops.Coalesce([self, *args]).to_expr()
 
-    def greatest(self, *args: ir.ValueExpr) -> ir.ValueExpr:
+    def greatest(self, *args: ir.Value) -> ir.Value:
         """Compute the largest value among the supplied arguments.
 
         Parameters
@@ -208,14 +208,14 @@ class AnyValue(ValueExpr):
 
         Returns
         -------
-        ValueExpr
+        Value
             Maximum of the passed arguments
         """
         import ibis.expr.operations as ops
 
         return ops.Greatest([self, *args]).to_expr()
 
-    def least(self, *args: ir.ValueExpr) -> ir.ValueExpr:
+    def least(self, *args: ir.Value) -> ir.Value:
         """Compute the smallest value among the supplied arguments.
 
         Parameters
@@ -225,7 +225,7 @@ class AnyValue(ValueExpr):
 
         Returns
         -------
-        ValueExpr
+        Value
             Minimum of the passed arguments
         """
         import ibis.expr.operations as ops
@@ -246,7 +246,7 @@ class AnyValue(ValueExpr):
 
         return ops.TypeOf(self).to_expr()
 
-    def fillna(self, fill_value: ScalarExpr) -> ValueExpr:
+    def fillna(self, fill_value: Scalar) -> Value:
         """Replace any null values with the indicated fill value.
 
         Parameters
@@ -263,14 +263,14 @@ class AnyValue(ValueExpr):
 
         Returns
         -------
-        ValueExpr
+        Value
             `self` filled with `fill_value` where it is `NA`
         """
         import ibis.expr.operations as ops
 
         return ops.IfNull(self, fill_value).to_expr()
 
-    def nullif(self, null_if_expr: ValueExpr) -> ValueExpr:
+    def nullif(self, null_if_expr: Value) -> Value:
         """Set values to null if they equal the values `null_if_expr`.
 
         Commonly use to avoid divide-by-zero problems by replacing zero with
@@ -283,7 +283,7 @@ class AnyValue(ValueExpr):
 
         Returns
         -------
-        ValueExpr
+        Value
             Value expression
         """
         import ibis.expr.operations as ops
@@ -292,8 +292,8 @@ class AnyValue(ValueExpr):
 
     def between(
         self,
-        lower: ValueExpr,
-        upper: ValueExpr,
+        lower: Value,
+        upper: Value,
     ) -> ir.BooleanValue:
         """Check if this expression is between `lower` and `upper`, inclusive.
 
@@ -316,7 +316,7 @@ class AnyValue(ValueExpr):
 
     def isin(
         self,
-        values: ValueExpr | Sequence[ValueExpr],
+        values: Value | Sequence[Value],
     ) -> ir.BooleanValue:
         """Check whether this expression's values are in `values`.
 
@@ -344,7 +344,7 @@ class AnyValue(ValueExpr):
 
     def notin(
         self,
-        values: ValueExpr | Sequence[ValueExpr],
+        values: Value | Sequence[Value],
     ) -> ir.BooleanValue:
         """Check whether this expression's values are not in `values`.
 
@@ -364,9 +364,9 @@ class AnyValue(ValueExpr):
 
     def substitute(
         self,
-        value: ValueExpr,
-        replacement: ValueExpr | None = None,
-        else_: ValueExpr | None = None,
+        value: Value,
+        replacement: Value | None = None,
+        else_: Value | None = None,
     ):
         """Replace one or more values in a value expression.
 
@@ -382,7 +382,7 @@ class AnyValue(ValueExpr):
 
         Returns
         -------
-        ValueExpr
+        Value
             Replaced values
         """
         expr = self.case()
@@ -394,7 +394,7 @@ class AnyValue(ValueExpr):
 
         return expr.else_(else_ if else_ is not None else self).end()
 
-    def over(self, window: win.Window) -> ValueExpr:
+    def over(self, window: win.Window) -> Value:
         """Construct a window expression.
 
         Parameters
@@ -404,7 +404,7 @@ class AnyValue(ValueExpr):
 
         Returns
         -------
-        ValueExpr
+        Value
             A window function expression
         """
         import ibis.expr.operations as ops
@@ -470,9 +470,9 @@ class AnyValue(ValueExpr):
 
     def cases(
         self,
-        case_result_pairs: Iterable[tuple[ir.BooleanValue, ValueExpr]],
-        default: ValueExpr | None = None,
-    ) -> ValueExpr:
+        case_result_pairs: Iterable[tuple[ir.BooleanValue, Value]],
+        default: Value | None = None,
+    ) -> Value:
         """Create a case expression in one shot.
 
         Parameters
@@ -484,7 +484,7 @@ class AnyValue(ValueExpr):
 
         Returns
         -------
-        ValueExpr
+        Value
             Value expression
         """
         builder = self.case()
@@ -498,7 +498,7 @@ class AnyValue(ValueExpr):
 
         return ops.ArrayCollect(self).to_expr()
 
-    def identical_to(self, other: ValueExpr) -> ir.BooleanValue:
+    def identical_to(self, other: Value) -> ir.BooleanValue:
         """Return whether this expression is identical to other.
 
         Corresponds to `IS NOT DISTINCT FROM` in SQL.
@@ -585,13 +585,13 @@ class AnyValue(ValueExpr):
 
 
 @public
-class AnyScalar(ScalarExpr, AnyValue):
+class AnyScalar(Scalar, AnyValue):
     pass  # noqa: E701,E302
 
 
 @public
-class AnyColumn(ColumnExpr, AnyValue):
-    def bottomk(self, k: int, by: ValueExpr | None = None) -> ir.TopKExpr:
+class AnyColumn(Column, AnyValue):
+    def bottomk(self, k: int, by: Value | None = None) -> ir.TopKExpr:
         raise NotImplementedError("bottomk is not implemented")
 
     def approx_nunique(
@@ -605,17 +605,17 @@ class AnyColumn(ColumnExpr, AnyValue):
     def approx_median(
         self,
         where: ir.BooleanValue | None = None,
-    ) -> ScalarExpr:
+    ) -> Scalar:
         import ibis.expr.operations as ops
 
         return ops.CMSMedian(self, where).to_expr().name("approx_median")
 
-    def max(self, where: ir.BooleanValue | None = None) -> ScalarExpr:
+    def max(self, where: ir.BooleanValue | None = None) -> Scalar:
         import ibis.expr.operations as ops
 
         return ops.Max(self, where).to_expr().name("max")
 
-    def min(self, where: ir.BooleanValue | None = None) -> ScalarExpr:
+    def min(self, where: ir.BooleanValue | None = None) -> Scalar:
         import ibis.expr.operations as ops
 
         return ops.Min(self, where).to_expr().name("min")
@@ -630,7 +630,7 @@ class AnyColumn(ColumnExpr, AnyValue):
     def topk(
         self,
         k: int,
-        by: ir.ValueExpr | None = None,
+        by: ir.Value | None = None,
     ) -> ir.TopKExpr:
         """Return a "top k" expression.
 
@@ -692,7 +692,7 @@ class AnyColumn(ColumnExpr, AnyValue):
         self,
         where: ir.BooleanValue | None = None,
         how: str | None = None,
-    ) -> ScalarExpr:
+    ) -> Scalar:
         """Select an arbitrary value in a column.
 
         Parameters
@@ -705,7 +705,7 @@ class AnyColumn(ColumnExpr, AnyValue):
 
         Returns
         -------
-        ScalarExpr
+        Scalar
             An expression
         """
         import ibis.expr.operations as ops
@@ -749,37 +749,37 @@ class AnyColumn(ColumnExpr, AnyValue):
 
         return base.group_by(expr).aggregate(metric)
 
-    def first(self) -> ColumnExpr:
+    def first(self) -> Column:
         import ibis.expr.operations as ops
 
         return ops.FirstValue(self).to_expr()
 
-    def last(self) -> ColumnExpr:
+    def last(self) -> Column:
         import ibis.expr.operations as ops
 
         return ops.LastValue(self).to_expr()
 
-    def rank(self) -> ColumnExpr:
+    def rank(self) -> Column:
         import ibis.expr.operations as ops
 
         return ops.MinRank(self).to_expr()
 
-    def dense_rank(self) -> ColumnExpr:
+    def dense_rank(self) -> Column:
         import ibis.expr.operations as ops
 
         return ops.DenseRank(self).to_expr()
 
-    def percent_rank(self) -> ColumnExpr:
+    def percent_rank(self) -> Column:
         import ibis.expr.operations as ops
 
         return ops.PercentRank(self).to_expr()
 
-    def cummin(self) -> ColumnExpr:
+    def cummin(self) -> Column:
         import ibis.expr.operations as ops
 
         return ops.CumulativeMin(self).to_expr()
 
-    def cummax(self) -> ColumnExpr:
+    def cummax(self) -> Column:
         import ibis.expr.operations as ops
 
         return ops.CumulativeMax(self).to_expr()
@@ -787,8 +787,8 @@ class AnyColumn(ColumnExpr, AnyValue):
     def lag(
         self,
         offset: int | ir.IntegerValue | None = None,
-        default: ValueExpr | None = None,
-    ) -> ColumnExpr:
+        default: Value | None = None,
+    ) -> Column:
         import ibis.expr.operations as ops
 
         return ops.Lag(self, offset, default).to_expr()
@@ -796,8 +796,8 @@ class AnyColumn(ColumnExpr, AnyValue):
     def lead(
         self,
         offset: int | ir.IntegerValue | None = None,
-        default: ValueExpr | None = None,
-    ) -> ColumnExpr:
+        default: Value | None = None,
+    ) -> Column:
         import ibis.expr.operations as ops
 
         return ops.Lead(self, offset, default).to_expr()
@@ -807,7 +807,7 @@ class AnyColumn(ColumnExpr, AnyValue):
 
         return ops.NTile(self, buckets).to_expr()
 
-    def nth(self, n: int | ir.IntegerValue) -> ColumnExpr:
+    def nth(self, n: int | ir.IntegerValue) -> Column:
         """Return the `n`th value over a window.
 
         Parameters
@@ -817,7 +817,7 @@ class AnyColumn(ColumnExpr, AnyValue):
 
         Returns
         -------
-        ColumnExpr
+        Column
             The nth value over a window
         """
         import ibis.expr.operations as ops
@@ -840,9 +840,17 @@ class NullColumn(AnyColumn, NullValue):
     pass  # noqa: E701,E302
 
 
-# TODO(kszucs): should remove the ColumnExpr base class?
+class List(Expr):
+    pass
+
+
+class ValueList(List, AnyValue):
+    pass
+
+
+# TODO(kszucs): should remove the Column base class?
 @public
-class ListExpr(ColumnExpr, AnyValue):
+class ListExpr(Column, AnyValue):
     @property
     def values(self):
         return self.op().values
@@ -886,7 +894,7 @@ def null():
 
 
 @public
-def literal(value: Any, type: dt.DataType | str | None = None) -> ScalarExpr:
+def literal(value: Any, type: dt.DataType | str | None = None) -> Scalar:
     """Create a scalar expression from a Python value.
 
     !!! tip "Use specific functions for arrays, structs and maps"
@@ -913,7 +921,7 @@ def literal(value: Any, type: dt.DataType | str | None = None) -> ScalarExpr:
 
     Returns
     -------
-    ScalarExpr
+    Scalar
         An expression representing a literal value
 
     Examples
@@ -981,3 +989,7 @@ def literal(value: Any, type: dt.DataType | str | None = None) -> ScalarExpr:
     else:
         value = dt._normalize(dtype, value)
         return ops.Literal(value, dtype=dtype).to_expr()
+
+
+# Aliases for backward compatibility
+public(ValueExpr=Expr, ScalarExpr=Scalar, ColumnExpr=Column)