+++ ibis/expr/types/groupby.py
@@ -117,7 +117,7 @@ class GroupedTableExpr:
         )
 
     def order_by(
-        self, expr: ir.ValueExpr | Iterable[ir.ValueExpr]
+        self, expr: ir.Value | Iterable[ir.Value]
     ) -> GroupedTableExpr:
         """Sort a grouped table expression by `expr`.
 
@@ -147,8 +147,8 @@ class GroupedTableExpr:
 
     def mutate(
         self,
-        exprs: ir.ValueExpr | Sequence[ir.ValueExpr] | None = None,
-        **kwds: ir.ValueExpr,
+        exprs: ir.Value | Sequence[ir.Value] | None = None,
+        **kwds: ir.Value,
     ):
         """Return a table projection with window functions applied.